requirements.txt
absl-py==0.15.0
adversarial-robustness-toolbox==1.2.0
alembic==1.7.6
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
auto-sklearn==0.14.0
autopage==0.5.0
backcall==0.2.0
beautifulsoup4==4.10.0
bitarray==2.4.0
boto3==1.21.15
botocore==1.24.15
Brotli==1.0.9
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
chainer==7.8.1
chardet==4.0.0
charset-normalizer==2.0.12
cleaningbenchmark==0.0.1
cleanlab==1.0
click==8.0.4
cliff==3.10.1
cloudpickle==2.0.0
cmaes==0.8.2
cmd2==2.4.0
colorlog==6.6.0
ConfigSpace==0.4.21
cycler==0.11.0
Cython==0.29.2
dask==2021.6.2
datawig==0.2.0
deap==1.3.1
debugpy==1.5.1
decorator==5.1.1
distributed==2021.6.2
eli5==0.11.0
emcee==3.1.1
entrypoints==0.4
errorgenerator==0.2
et-xmlfile==1.1.0
executing==0.8.3
filelock==3.6.0
flatbuffers==1.12
fonttools==4.29.1
fsspec==2022.2.0
future==0.18.2
gast==0.4.0
gensim==3.8.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.8.4
greenlet==1.1.2
grpcio==1.34.1
h5py==3.1.0
HeapDict==1.0.1
idna==3.3
ijson==3.1.4
imbalanced-learn==0.8.0
imblearn==0.0
importlib-metadata==4.11.2
importlib-resources==5.4.0
impyute==0.0.8
ipykernel==6.9.1
ipython==8.1.1
isodate==0.6.1
jarowinkler==1.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonlines==3.0.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Levenshtein==0.18.1
liac-arff==2.5.0
linear-tsv==1.1.0
locket==0.2.1
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
missingpy==0.2.0
mlxtend==0.18.0
msgpack==1.0.3
multivolumefile==0.2.3
mwparserfromhell==0.6.4
mxnet==1.4.0
nest-asyncio==1.5.4
networkx==2.7.1
numpy==1.21.2
oauthlib==3.2.0
openpyxl==3.0.9
opt-einsum==3.3.0
optuna==2.7.0
packaging==21.3
pandas==1.3.3
parso==0.8.3
partd==1.2.0
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prettytable==3.2.0
probableparsing==0.0.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-entitymatching==0.4.0
py-stringmatching==0.4.2
py-stringsimjoin==0.3.2
py7zr==0.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==0.5.0
pybind11==2.9.1
pybloom-live==3.1.0
pycryptodomex==3.14.1
Pygments==2.11.2
pyitlib==0.2.2
pynisher==0.6.4
pyparsing==3.0.7
pyperclip==1.8.2
pyppmd==0.17.4
PyPrind==2.11.3
pyrfr==0.8.2
pyrsistent==0.18.1
python-crfsuite==0.9.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
pyzstd==0.15.2
raha==1.25
rapidfuzz==2.0.6
regex==2021.4.4
rein==0.1.0
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==2.0.0
rsa==4.8
s3transfer==0.5.2
scikit-learn==0.24.2
scikit-sparse==0.4.5
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
smac==1.2
smart-open==5.2.1
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.2
stack-data==0.2.0
stevedore==3.5.0
stopit==1.1.2
tableschema==1.20.2
tableschema-sql==1.3.1
tabulate==0.8.9
tabulator==1.53.5
tblib==1.7.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
texttable==1.6.4
threadpoolctl==3.1.0
toolz==0.11.2
torch==1.10.2+cpu
torchaudio==0.10.2+cpu
torchvision==0.11.3+cpu
tornado==6.1
TPOT==0.11.7
tqdm==4.63.0
traitlets==5.1.1
typing==3.6.6
typing-extensions==4.1.1
unicodecsv==0.14.1
update-checker==0.18.0
urllib3==1.26.8
usaddress==0.5.10
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.12.1
xgboost==1.3.3
xlrd==2.0.1
zict==2.1.0
zipfile-deflate64==0.2.0
zipp==3.7.0
